Almost ready for release (95%)
Essentially ready for release.
+ IO should be working
- automatically load from default location (same location the app is
opened from)
- import .xml. export in .txt. The program does export .xml as well, but
the file is the default file used when the program starts up and is
found in the same directory as the application
- right click context menus on rows for same button options (add, edit,
delete, and... Searching online)
- Menu bar used to access import and export options. Refresh enabled but
not needed, used when testing program before. About menu for the lolz
- added new Film attribute - Comments. User can put comments on each
film and of course everything is saved and displayed
- in the Add/Edit window for films there is a useless Date picker that
is not used by the application yet. Experimental
--------------------------
todo:
- add functional date attribute to film
- add more attributes to be displayed
- maybe more features?

same